+++ src/SummaryTables.jl
@@ -19,6 +19,7 @@ import NaturalSort
 import WriteDocx
 import SHA
 
+export simple_table
 export table_one
 export listingtable
 export summarytable
@@ -40,6 +41,7 @@ include("infrastructure/cells.jl")
 include("infrastructure/special_cell_values.jl")
 include("infrastructure/tables.jl")
 
+include("table_functions/simple_table.jl")
 include("table_functions/table_one.jl")
 include("table_functions/listingtable.jl")
 include("table_functions/summarytable.jl")